+++ src/main.cpp
@@ -104,8 +104,8 @@ int main( void )
         glViewport(0, 0, width, height);
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
         
-        double currentTime = glfwGetTime();
-        double delta = currentTime - lastTime;
+        float currentTime = glfwGetTime();
+        float delta = currentTime - lastTime;
         lastTime = currentTime;
